refactor: extract search providers from deep_research.py (Phase 2.3)
The 3,681-LOC `maverick_mcp/agents/deep_research.py` is now 3,032 LOC
after lifting the search-provider strategies into a focused subpackage.
Public API is preserved via re-exports.

New `maverick_mcp/agents/research/providers/`:

| File | Responsibility |
|------|-----------------|
| `base.py` | `WebSearchProvider` — health tracking, timeout policy, multi-provider aggregation |
| `exa.py` | `ExaSearchProvider` — Exa-specific strategies, financial-relevance scoring |
| `tavily.py` | `TavilySearchProvider` — Tavily wrapper with domain filtering |
| `__init__.py` | Re-exports + `get_cached_search_provider()` and `_search_provider_cache` |

`deep_research.py` re-exports the same five symbols at the legacy path
so existing call sites (`api/routers/research.py`,
`api/routers/agents.py`, `agents/optimized_research.py`) and tests
(notably `tests/test_exa_research_integration.py` which patches
`maverick_mcp.agents.deep_research.get_cached_search_provider`) keep
working without edits. Marked `# noqa: F401` on the import block and
listed the names in `__all__` so unused-import linters don't strip
them.

7 new characterization tests in
`tests/agents/test_deep_research_characterization.py` lock in the
contract:
- legacy and new paths re-export the same class objects
- `RESEARCH_DEPTH_LEVELS` and `PERSONA_RESEARCH_FOCUS` shapes are
  stable
- `WebSearchProvider` health lifecycle (record_failure ×7 → unhealthy,
  record_success → healthy)
- `_calculate_timeout` scales with query word count

Out of scope (deferred to a follow-up): the state-machine extraction
into `agents/research/state.py`. The state transitions are woven
through `DeepResearchAgent` itself rather than living as a separable
function — that extraction needs its own characterization-driven PR
to avoid regressing the LangGraph workflow.

Result: 892 unit tests pass (up from 885 — 7 new characterization
cases). Ruff clean. Strict-zone ty check still 0 diagnostics.
