+++ maverick_mcp/agents/deep_research.py
@@ -10,7 +10,6 @@ from __future__ import annotations
 import asyncio
 import json
 import logging
-from collections.abc import Iterable
 from datetime import UTC, datetime
 from typing import Any
 from uuid import uuid4
@@ -23,11 +22,21 @@ from langgraph.graph import END, START, StateGraph  # type: ignore[import-untype
 from langgraph.types import Command  # type: ignore[import-untyped]
 
 from maverick_mcp.agents.base import PersonaAwareAgent
-from maverick_mcp.agents.circuit_breaker import circuit_manager
-from maverick_mcp.config.settings import get_settings
-from maverick_mcp.exceptions import (
-    WebSearchError,
+
+# Provider classes were extracted to maverick_mcp.agents.research.providers
+# (Phase 2.3). Re-exports below preserve the public import surface — call
+# sites and tests still write `from maverick_mcp.agents.deep_research
+# import ExaSearchProvider`. Listed in `__all__` (bottom of file) so
+# unused-import linters don't strip them.
+from maverick_mcp.agents.research.providers import (  # noqa: F401
+    ExaSearchProvider,
+    TavilySearchProvider,
+    WebSearchProvider,
+    _search_provider_cache,
+    get_cached_search_provider,
 )
+from maverick_mcp.config.settings import get_settings
+from maverick_mcp.data.vector_store import ResearchResult, get_vector_store
 from maverick_mcp.memory.checkpointer import get_persistent_checkpointer
 from maverick_mcp.memory.stores import ConversationStore
 from maverick_mcp.utils.orchestration_logging import (
@@ -37,45 +46,10 @@ from maverick_mcp.utils.orchestration_logging import (
     log_performance_metrics,
     log_synthesis_operation,
 )
-
-try:  # pragma: no cover - optional dependency
-    from tavily import TavilyClient  # type: ignore[import-not-found]
-except ImportError:  # pragma: no cover
-    TavilyClient = None  # type: ignore[assignment]
-
-# Import moved to avoid circular dependency - will import where needed
-from maverick_mcp.data.vector_store import ResearchResult, get_vector_store
 from maverick_mcp.workflows.state import DeepResearchState
 
 logger = logging.getLogger(__name__)
 settings = get_settings()
-
-# Global search provider cache and connection manager
-_search_provider_cache: dict[str, Any] = {}
-
-
-async def get_cached_search_provider(exa_api_key: str | None = None) -> Any | None:
-    """Get cached Exa search provider to avoid repeated initialization delays."""
-    cache_key = f"exa:{exa_api_key is not None}"
-
-    if cache_key in _search_provider_cache:
-        return _search_provider_cache[cache_key]
-
-    logger.info("Initializing Exa search provider")
-    provider = None
-
-    # Initialize Exa provider with caching
-    if exa_api_key:
-        try:
-            provider = ExaSearchProvider(exa_api_key)
-            logger.info("Initialized Exa search provider")
-            # Cache the provider
-            _search_provider_cache[cache_key] = provider
-        except ImportError as e:
-            logger.warning(f"Failed to initialize Exa provider: {e}")
-
-    return provider
-
 
 # Research depth levels optimized for quick searches
 RESEARCH_DEPTH_LEVELS = {
@@ -156,649 +130,6 @@ PERSONA_RESEARCH_FOCUS = {
         "time_horizon": "intraday to weekly",
     },
 }
-
-
-class WebSearchProvider:
-    """Base class for web search providers with early abort mechanism."""
-
-    def __init__(self, api_key: str):
-        self.api_key = api_key
-        self.rate_limiter = None  # Implement rate limiting
-        self._failure_count = 0
-        self._max_failures = 3  # Abort after 3 consecutive failures
-        self._is_healthy = True
-        self.settings = get_settings()
-        self.logger = logging.getLogger(f"{__name__}.{self.__class__.__name__}")
-
-    def _calculate_timeout(
-        self, query: str, timeout_budget: float | None = None
-    ) -> float:
-        """Calculate generous timeout for thorough research operations."""
-        query_words = len(query.split())
-
-        # Generous timeout calculation for thorough search operations
-        if query_words <= 3:
-            base_timeout = 30.0  # Simple queries - 30s for thorough results
-        elif query_words <= 8:
-            base_timeout = 45.0  # Standard queries - 45s for comprehensive search
-        else:
-            base_timeout = 60.0  # Complex queries - 60s for exhaustive search
-
-        # Apply budget constraints if available
-        if timeout_budget and timeout_budget > 0:
-            # Use generous portion of available budget per search operation
-            budget_timeout = max(
-                timeout_budget * 0.6, 30.0
-            )  # At least 30s, use 60% of budget
-            calculated_timeout = min(base_timeout, budget_timeout)
-
-            # Ensure minimum timeout (at least 30s for thorough search)
-            calculated_timeout = max(calculated_timeout, 30.0)
-        else:
-            calculated_timeout = base_timeout
-
-        # Final timeout with generous minimum for thorough search
-        final_timeout = max(calculated_timeout, 30.0)
-
-        return final_timeout
-
-    def _record_failure(self, error_type: str = "unknown") -> None:
-        """Record a search failure and check if provider should be disabled."""
-        self._failure_count += 1
-
-        # Use separate thresholds for timeout vs other failures
-        timeout_threshold = getattr(
-            self.settings.performance, "search_timeout_failure_threshold", 12
-        )
-
-        # Much more tolerant of timeout failures - they may be due to network/complexity
-        if error_type == "timeout" and self._failure_count >= timeout_threshold:
-            self._is_healthy = False
-            logger.warning(
-                f"Search provider {self.__class__.__name__} disabled after "
-                f"{self._failure_count} consecutive timeout failures (threshold: {timeout_threshold})"
-            )
-        elif error_type != "timeout" and self._failure_count >= self._max_failures * 2:
-            # Be more lenient for non-timeout failures (2x threshold)
-            self._is_healthy = False
-            logger.warning(
-                f"Search provider {self.__class__.__name__} disabled after "
-                f"{self._failure_count} total non-timeout failures"
-            )
-
-        logger.debug(
-            f"Provider {self.__class__.__name__} failure recorded: "
-            f"type={error_type}, count={self._failure_count}, healthy={self._is_healthy}"
-        )
-
-    def _record_success(self) -> None:
-        """Record a successful search and reset failure count."""
-        if self._failure_count > 0:
-            logger.info(
-                f"Search provider {self.__class__.__name__} recovered after "
-                f"{self._failure_count} failures"
-            )
-        self._failure_count = 0
-        self._is_healthy = True
-
-    def is_healthy(self) -> bool:
-        """Check if provider is healthy and should be used."""
-        return self._is_healthy
-
-    async def search(
-        self, query: str, num_results: int = 10, timeout_budget: float | None = None
-    ) -> list[dict[str, Any]]:
-        """Perform web search and return results."""
-        raise NotImplementedError
-
-    async def get_content(self, url: str) -> dict[str, Any]:
-        """Extract content from URL."""
-        raise NotImplementedError
-
-    async def search_multiple_providers(
-        self,
-        queries: list[str],
-        providers: list[str] | None = None,
-        max_results_per_query: int = 5,
-    ) -> dict[str, list[dict[str, Any]]]:
-        """Search using multiple providers and return aggregated results."""
-        providers = providers or ["exa"]  # Default to available providers
-        results = {}
-
-        for provider_name in providers:
-            provider_results = []
-            for query in queries:
-                try:
-                    query_results = await self.search(query, max_results_per_query)
-
-                    provider_results.extend(query_results or [])
-                except Exception as e:
-                    self.logger.warning(
-                        f"Search failed for provider {provider_name}, query '{query}': {e}"
-                    )
-                    continue
-
-            results[provider_name] = provider_results
-
-        return results
-
-    def _timeframe_to_date(self, timeframe: str) -> str | None:
-        """Convert timeframe string to date string."""
-        from datetime import datetime, timedelta
-
-        now = datetime.now()
-
-        if timeframe == "1d":
-            date = now - timedelta(days=1)
-        elif timeframe == "1w":
-            date = now - timedelta(weeks=1)
-        elif timeframe == "1m":
-            date = now - timedelta(days=30)
-        else:
-            # Invalid or unsupported timeframe, return None
-            return None
-
-        return date.strftime("%Y-%m-%d")
-
-
-class ExaSearchProvider(WebSearchProvider):
-    """Exa search provider for comprehensive web search using MCP tools with financial optimization."""
-
-    def __init__(self, api_key: str):
-        super().__init__(api_key)
-        # Store the API key for verification
-        self._api_key_verified = bool(api_key)
-
-        # Financial-specific domain preferences for better results
-        self.financial_domains = [
-            "sec.gov",
-            "edgar.sec.gov",
-            "investor.gov",
-            "bloomberg.com",
-            "reuters.com",
-            "wsj.com",
-            "ft.com",
-            "marketwatch.com",
-            "yahoo.com/finance",
-            "finance.yahoo.com",
-            "morningstar.com",
-            "fool.com",
-            "seekingalpha.com",
-            "investopedia.com",
-            "barrons.com",
-            "cnbc.com",
-            "nasdaq.com",
-            "nyse.com",
-            "finra.org",
-            "federalreserve.gov",
-            "treasury.gov",
-            "bls.gov",
-        ]
-
-        # Domains to exclude for financial searches
-        self.excluded_domains = [
-            "facebook.com",
-            "twitter.com",
-            "x.com",
-            "instagram.com",
-            "tiktok.com",
-            "reddit.com",
-            "pinterest.com",
-            "linkedin.com",
-            "youtube.com",
-            "wikipedia.org",
-        ]
-
-        logger.info("Initialized ExaSearchProvider with financial optimization")
-
-    async def search(
-        self, query: str, num_results: int = 10, timeout_budget: float | None = None
-    ) -> list[dict[str, Any]]:
-        """Search using Exa via async client for comprehensive web results with adaptive timeout."""
-        return await self._search_with_strategy(
-            query, num_results, timeout_budget, "auto"
-        )
-
-    async def search_financial(
-        self,
-        query: str,
-        num_results: int = 10,
-        timeout_budget: float | None = None,
-        strategy: str = "hybrid",
-    ) -> list[dict[str, Any]]:
-        """
-        Enhanced financial search with optimized queries and domain targeting.
-
-        Args:
-            query: Search query
-            num_results: Number of results to return
-            timeout_budget: Timeout budget in seconds
-            strategy: Search strategy - 'hybrid', 'authoritative', 'comprehensive', or 'auto'
-        """
-        return await self._search_with_strategy(
-            query, num_results, timeout_budget, strategy
-        )
-
-    async def _search_with_strategy(
-        self, query: str, num_results: int, timeout_budget: float | None, strategy: str
-    ) -> list[dict[str, Any]]:
-        """Internal method to handle different search strategies."""
-
-        # Check provider health before attempting search
-        if not self.is_healthy():
-            logger.warning("Exa provider is unhealthy - skipping search")
-            raise WebSearchError("Exa provider disabled due to repeated failures")
-
-        # Calculate adaptive timeout
-        search_timeout = self._calculate_timeout(query, timeout_budget)
-
-        try:
-            # Use search-specific circuit breaker settings (more tolerant)
-            circuit_breaker = await circuit_manager.get_or_create(
-                "exa_search",
-                failure_threshold=getattr(
-                    self.settings.performance,
-                    "search_circuit_breaker_failure_threshold",
-                    8,
-                ),
-                recovery_timeout=getattr(
-                    self.settings.performance,
-                    "search_circuit_breaker_recovery_timeout",
-                    30,
-                ),
-            )
-
-            async def _search():
-                # Use the async exa-py library for web search
-                try:
-                    from exa_py import AsyncExa
-
-                    # Initialize AsyncExa client with API key
-                    async_exa_client = AsyncExa(api_key=self.api_key)
-
-                    # Configure search parameters based on strategy
-                    search_params = self._get_search_params(
-                        query, num_results, strategy
-                    )
-
-                    # Call Exa search with optimized parameters
-                    exa_response = await async_exa_client.search_and_contents(
-                        **search_params
-                    )
-
-                    # Convert Exa response to standard format with enhanced metadata
-                    results = []
-                    if exa_response and hasattr(exa_response, "results"):
-                        for result in exa_response.results:
-                            # Enhanced result processing with financial relevance scoring
-                            financial_relevance = self._calculate_financial_relevance(
-                                result
-                            )
-
-                            results.append(
-                                {
-                                    "url": result.url or "",
-                                    "title": result.title or "No Title",
-                                    "content": (result.text or "")[:2000],
-                                    "raw_content": (result.text or "")[
-                                        :5000
-                                    ],  # Increased for financial content
-                                    "published_date": result.published_date or "",
-                                    "score": result.score
-                                    if hasattr(result, "score")
-                                    and result.score is not None
-                                    else 0.7,
-                                    "financial_relevance": financial_relevance,
-                                    "provider": "exa",
-                                    "author": result.author
-                                    if hasattr(result, "author")
-                                    and result.author is not None
-                                    else "",
-                                    "domain": self._extract_domain(result.url or ""),
-                                    "is_authoritative": self._is_authoritative_source(
-                                        result.url or ""
-                                    ),
-                                }
-                            )
-
-                    # Sort results by financial relevance and score
-                    results.sort(
-                        key=lambda x: (x["financial_relevance"], x["score"]),
-                        reverse=True,
-                    )
-                    return results
-
-                except ImportError:
-                    logger.error("exa-py library not available - cannot perform search")
-                    raise WebSearchError(
-                        "exa-py library required for ExaSearchProvider"
-                    )
-                except Exception as e:
-                    logger.error(f"Error calling Exa API: {e}")
-                    raise e
-
-            # Use adaptive timeout based on query complexity and budget
-            result = await asyncio.wait_for(
-                circuit_breaker.call(_search), timeout=search_timeout
-            )
-            self._record_success()  # Record successful search
-            logger.debug(
-                f"Exa search completed in {search_timeout:.1f}s timeout window"
-            )
-            return result
-
-        except TimeoutError:
-            self._record_failure("timeout")  # Record timeout as specific failure type
-            query_snippet = query[:100] + ("..." if len(query) > 100 else "")
-            logger.error(
-                f"Exa search timeout after {search_timeout:.1f} seconds (failure #{self._failure_count}) "
-                f"for query: '{query_snippet}'"
-            )
-            raise WebSearchError(
-                f"Exa search timed out after {search_timeout:.1f} seconds"
-            )
-        except Exception as e:
-            self._record_failure("error")  # Record non-timeout failure
-            logger.error(f"Exa search error (failure #{self._failure_count}): {e}")
-            raise WebSearchError(f"Exa search failed: {str(e)}")
-
-    def _get_search_params(
-        self, query: str, num_results: int, strategy: str
-    ) -> dict[str, Any]:
-        """
-        Generate optimized search parameters based on strategy and query type.
-
-        Args:
-            query: Search query
-            num_results: Number of results
-            strategy: Search strategy
-
-        Returns:
-            Dictionary of search parameters for Exa API
-        """
-        # Base parameters
-        params = {
-            "query": query,
-            "num_results": num_results,
-            "text": {"max_characters": 5000},  # Increased for financial content
-        }
-
-        # Strategy-specific optimizations
-        if strategy == "authoritative":
-            # Focus on authoritative financial sources
-            # Note: Exa API doesn't allow both include_domains and exclude_domains with content
-            params.update(
-                {
-                    "include_domains": self.financial_domains[
-                        :10
-                    ],  # Top authoritative sources
-                    "type": "auto",  # Let Exa decide neural vs keyword
-                    "start_published_date": "2020-01-01",  # Recent financial data
-                }
-            )
-
-        elif strategy == "comprehensive":
-            # Broad search across all financial sources
-            params.update(
-                {
-                    "exclude_domains": self.excluded_domains,
-                    "type": "neural",  # Better for comprehensive understanding
-                    "start_published_date": "2018-01-01",  # Broader historical context
-                }
-            )
-
-        elif strategy == "hybrid":
-            # Balanced approach with domain preferences
-            params.update(
-                {
-                    "exclude_domains": self.excluded_domains,
-                    "type": "auto",  # Hybrid neural/keyword approach
-                    "start_published_date": "2019-01-01",
-                    # Use domain weighting rather than strict inclusion
-                }
-            )
-
-        else:  # "auto" or default
-            # Standard search with basic optimizations
-            params.update(
-                {
-                    "exclude_domains": self.excluded_domains[:5],  # Basic exclusions
-                    "type": "auto",
-                }
-            )
-
-        # Add financial-specific query enhancements
-        enhanced_query = self._enhance_financial_query(query)
-        if enhanced_query != query:
-            params["query"] = enhanced_query
-
-        return params
-
-    def _enhance_financial_query(self, query: str) -> str:
-        """
-        Enhance queries with financial context and terminology.
-
-        Args:
-            query: Original search query
-
-        Returns:
-            Enhanced query with financial context
-        """
-        # Financial keywords that improve search quality
-        financial_terms = {
-            "earnings",
-            "revenue",
-            "profit",
-            "loss",
-            "financial",
-            "quarterly",
-            "annual",
-            "SEC",
-            "10-K",
-            "10-Q",
-            "balance sheet",
-            "income statement",
-            "cash flow",
-            "dividend",
-            "stock",
-            "share",
-            "market cap",
-            "valuation",
-        }
-
-        query_lower = query.lower()
-
-        # Check if query already contains financial terms
-        has_financial_context = any(term in query_lower for term in financial_terms)
-
-        # Add context for company/stock queries
-        if not has_financial_context:
-            # Detect if it's a company or stock symbol query
-            if any(
-                indicator in query_lower
-                for indicator in ["company", "corp", "inc", "$", "stock"]
-            ):
-                return f"{query} financial analysis earnings revenue"
-            elif len(query.split()) <= 3 and query.isupper():  # Likely stock symbol
-                return f"{query} stock financial performance earnings"
-            elif "analysis" in query_lower or "research" in query_lower:
-                return f"{query} financial data SEC filings"
-
-        return query
-
-    def _calculate_financial_relevance(self, result) -> float:
-        """
-        Calculate financial relevance score for a search result.
-
-        Args:
-            result: Exa search result object
-
-        Returns:
-            Financial relevance score (0.0 to 1.0)
-        """
-        score = 0.0
-
-        # Domain-based scoring
-        domain = self._extract_domain(result.url)
-        if domain in self.financial_domains:
-            if domain in ["sec.gov", "edgar.sec.gov", "federalreserve.gov"]:
-                score += 0.4  # Highest authority
-            elif domain in ["bloomberg.com", "reuters.com", "wsj.com", "ft.com"]:
-                score += 0.3  # High-quality financial news
-            else:
-                score += 0.2  # Other financial sources
-
-        # Content-based scoring
-        if hasattr(result, "text") and result.text:
-            text_lower = result.text.lower()
-
-            # Financial terminology scoring
-            financial_keywords = [
-                "earnings",
-                "revenue",
-                "profit",
-                "financial",
-                "quarterly",
-                "annual",
-                "sec filing",
-                "10-k",
-                "10-q",
-                "balance sheet",
-                "income statement",
-                "cash flow",
-                "dividend",
-                "market cap",
-                "valuation",
-                "analyst",
-                "forecast",
-                "guidance",
-                "ebitda",
-                "eps",
-                "pe ratio",
-            ]
-
-            keyword_matches = sum(
-                1 for keyword in financial_keywords if keyword in text_lower
-            )
-            score += min(keyword_matches * 0.05, 0.3)  # Max 0.3 from keywords
-
-        # Title-based scoring
-        if hasattr(result, "title") and result.title:
-            title_lower = result.title.lower()
-            if any(
-                term in title_lower
-                for term in ["financial", "earnings", "quarterly", "annual", "sec"]
-            ):
-                score += 0.1
-
-        # Recency bonus for financial data
-        if hasattr(result, "published_date") and result.published_date:
-            try:
-                from datetime import datetime
-
-                # Handle different date formats
-                date_str = str(result.published_date)
-                if date_str and date_str != "":
-                    # Handle ISO format with Z
-                    if date_str.endswith("Z"):
-                        date_str = date_str.replace("Z", "+00:00")
-
-                    pub_date = datetime.fromisoformat(date_str)
-                    days_old = (datetime.now(UTC) - pub_date).days
-
-                    if days_old <= 30:
-                        score += 0.1  # Recent data bonus
-                    elif days_old <= 90:
-                        score += 0.05  # Somewhat recent bonus
-            except (ValueError, AttributeError, TypeError):
-                pass  # Skip if date parsing fails
-
-        return min(score, 1.0)  # Cap at 1.0
-
-    def _extract_domain(self, url: str) -> str:
-        """Extract domain from URL."""
-        try:
-            from urllib.parse import urlparse
-
-            return urlparse(url).netloc.lower().replace("www.", "")
-        except Exception:
-            return ""
-
-    def _is_authoritative_source(self, url: str) -> bool:
-        """Check if URL is from an authoritative financial source."""
-        domain = self._extract_domain(url)
-        authoritative_domains = [
-            "sec.gov",
-            "edgar.sec.gov",
-            "federalreserve.gov",
-            "treasury.gov",
-            "bloomberg.com",
-            "reuters.com",
-            "wsj.com",
-            "ft.com",
-        ]
-        return domain in authoritative_domains
-
-
-class TavilySearchProvider(WebSearchProvider):
-    """Tavily search provider with sensible filtering for financial research."""
-
-    def __init__(self, api_key: str):
-        super().__init__(api_key)
-        self.excluded_domains = {
-            "facebook.com",
-            "twitter.com",
-            "x.com",
-            "instagram.com",
-            "reddit.com",
-        }
-
-    async def search(
-        self, query: str, num_results: int = 10, timeout_budget: float | None = None
-    ) -> list[dict[str, Any]]:
-        if not self.is_healthy():
-            raise WebSearchError("Tavily provider disabled due to repeated failures")
-
-        timeout = self._calculate_timeout(query, timeout_budget)
-        circuit_breaker = await circuit_manager.get_or_create(
-            "tavily_search",
-            failure_threshold=8,
-            recovery_timeout=30,
-        )
-
-        async def _search() -> list[dict[str, Any]]:
-            if TavilyClient is None:
-                raise ImportError("tavily package is required for TavilySearchProvider")
-
-            client = TavilyClient(api_key=self.api_key)
-            response = await asyncio.get_event_loop().run_in_executor(
-                None,
-                lambda: client.search(query=query, max_results=num_results),
-            )
-            return self._process_results(response.get("results", []))
-
-        return await circuit_breaker.call(_search, timeout=timeout)
-
-    def _process_results(
-        self, results: Iterable[dict[str, Any]]
-    ) -> list[dict[str, Any]]:
-        processed: list[dict[str, Any]] = []
-        for item in results:
-            url = item.get("url", "")
-            if any(domain in url for domain in self.excluded_domains):
-                continue
-            processed.append(
-                {
-                    "url": url,
-                    "title": item.get("title"),
-                    "content": item.get("content") or item.get("raw_content", ""),
-                    "raw_content": item.get("raw_content"),
-                    "published_date": item.get("published_date"),
-                    "score": item.get("score", 0.0),
-                    "provider": "tavily",
-                }
-            )
-        return processed
 
 
 class ContentAnalyzer:
@@ -3679,3 +3010,21 @@ class CompetitiveResearchAgent(BaseSubagent):
             return 0.5
         credibility_scores = [r.get("credibility_score", 0.5) for r in results]
         return sum(credibility_scores) / len(credibility_scores)
+
+
+__all__ = [
+    "BaseSubagent",
+    "CompetitiveResearchAgent",
+    "ContentAnalyzer",
+    "DeepResearchAgent",
+    "ExaSearchProvider",
+    "FundamentalResearchAgent",
+    "PERSONA_RESEARCH_FOCUS",
+    "RESEARCH_DEPTH_LEVELS",
+    "SentimentResearchAgent",
+    "TavilySearchProvider",
+    "TechnicalResearchAgent",
+    "WebSearchProvider",
+    "_search_provider_cache",
+    "get_cached_search_provider",
+]
