+++ maverick_mcp/agents/research/__init__.py
@@ -1,0 +1,7 @@
+"""Research subpackage extracted from `agents/deep_research.py`.
+
+This subpackage holds the seams pulled out by the Phase 2.3 refactor —
+search-provider strategies in :mod:`.providers`, with state-machine
+extraction tracked as a follow-up. See
+``docs/superpowers/plans/2026-Q3-maverick-roadmap-v2.md`` for context.
+"""
